Stories moved (#26)
* Button components and some changes on webpack builder

* buttons, input components

* finished components

* style scroll on chrome

* changed all stories inside their components

* Composer component
