+++ src/components/Composer/style.scss
@@ -1,0 +1,32 @@
+@import '~styles/colors';
+@import '~styles/variables';
+
+
+.composer {
+    border: $default-border solid $color-text-light;
+    border-radius: $default-border-radius;
+    padding: 12px;
+    display: flex;
+    align-items: flex-end;
+}
+
+.icon {
+    padding: 0 12px;
+}
+
+.input {
+    font-size: 14px;
+    font-weight: 500;
+    border:none;
+    outline:none;
+    max-height: 150px;
+    flex: 1 1 auto;
+    resize: none;
+    word-wrap: break-word;
+    overflow-y: auto;
+    &:empty:before {
+        content: attr(placeholder);
+        display: block; /* For Firefox */
+        color: $color-text-light;
+    }
+}
