+++ src/style/colors.scss
@@ -1,0 +1,14 @@
+$color-text-dark: #2F343D;
+$color-text-grey: #9EA2A8;
+$color-text-light: #CBCED1;
+
+$bg-color-dark: #2F343D;
+$bg-color-grey: #E1E5E8; // TODO NAO TEM ISSO NA PALHETA
+$bg-color-light: #F7F8FA;
+$bg-color-white: #FFFFFF;
+
+$color-dark-blue: #175CC4;
+$color-blue: #1D74F5;
+
+$color-dark-red: #E0364D;
+$color-red: #F5455C;
