+++ src/style/helpers.scss
@@ -1,0 +1,21 @@
+$breakpoints: (
+  'small':  ( min-width:  767px ),
+  'medium': ( min-width:  992px ),
+  'large':  ( min-width: 1200px )
+) !default;
+
+@mixin md($breakpoint) {
+  // If the key exists in the map
+  @if map-has-key($breakpoints, $breakpoint) {
+    // Prints a media query based on the value
+    @media (#{inspect(map-get($breakpoints, $breakpoint))}) {
+      @content;
+    }
+  }
+
+  // If the key doesn't exist in the map
+  @else {
+    @warn "Unfortunately, there are no value to `#{$breakpoint}` breakpoints. "
+        + "Available breakpoints are: #{map-keys($breakpoints)}.";
+  }
+}
