+++ src/style/variables.scss
@@ -1,0 +1,13 @@
+
+$default-border: 2px;
+$default-border-radius: 2px;
+
+$form-item-height: 44px;
+$form-item-height-default: $form-item-height;
+$form-item-height-small: 30px;
+$form-item-margin-bottom: 16px;
+
+$form-input-label-font-size: 12px;
+$form-input-description-font-size: 12px;
+
+$default-time-animation: .3s;
