+++ src/styles/index.scss
@@ -1,4 +1,5 @@
 @import 'grid';
+@import 'colors';
 
 html, body {
 	height: 100%;
@@ -15,6 +16,20 @@ html, body {
 
 * {
 	box-sizing: border-box;
+}
+
+*::-webkit-scrollbar {
+  width: 10px;
+
+  &-thumb {
+	background: $color-text-grey;
+	border-radius: 5px;
+  }
+
+  &-track {
+	background: $color-text-light;
+	border-radius: 5px;
+  }
 }
 
 #app {
