+++ stories/index.js
@@ -1,7 +1,13 @@
 import { h } from 'preact';
+import { setOptions } from '@storybook/addon-options';
+import '../src/styles';
 
-import '../src/style';
-import './Button';
-import './Header';
-import './Footer';
-import './Input';
+function importAll(r) {
+	r.keys().forEach((key) => r(key));
+}
+
+setOptions({
+	hierarchySeparator: '\\.',
+});
+
+importAll(require.context('../src', true, /story$/));
